requirements.txt
langchain-community
langchain-core
langchain-openai
chromadb
openai
pypdf
python-dotenv